refactor logging I: add interface and move some files
